♻️ FAQ뷰 코드수정완료
최신 develop바탕으로 코드 리팩터토링, 뷰 분리완료
