+++ ttakkeun/ttakkeun/DataModelGroup/QnaJson/CategoryQnaJsonData.json
@@ -200,52 +200,52 @@
         "answer": "발톱이 뽑히지 않도록 주의해주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 자주 아파요.",
         "answer": "이빨 상태를 정기적으로 체크하고 양치질을 도와주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 흔들려요.",
         "answer": "이빨이 건강한지 확인해주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 변색됐어요.",
         "answer": "치아 청소를 주기적으로 해주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨에 이가 끼었어요.",
         "answer": "칫솔질을 도와주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨에 구멍이 생겼어요.",
         "answer": "충치 여부를 확인해주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 깨졌어요.",
         "answer": "동물병원에 방문하세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 너무 길어요.",
         "answer": "정기적인 치아 다듬기를 해주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 너무 짧아요.",
         "answer": "이빨 건강을 체크해주세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨에서 피가 나요.",
         "answer": "즉시 동물병원에 가세요."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 삐뚤빼뚤해요.",
         "answer": "치아 교정이 필요한지 확인해주세요."
     }
