+++ ttakkeun/ttakkeun/DataModelGroup/QnaJson/TopTenQnaJsonData.json
@@ -20,7 +20,7 @@
         "answer": "발톱은 주기적으로 깎아주는 것이 좋습니다. 너무 길어지면 걸을 때 불편함을 줄 수 있고, 심한 경우 발톱이 말려 들어가 피부를 다칠 수 있습니다."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 자주 아파요.",
         "answer": "치아 문제가 있는 경우 통증이 발생할 수 있습니다. 치석 제거와 같은 치아 관리를 주기적으로 해주고, 필요한 경우 수의사와 상담하세요."
     },
@@ -45,7 +45,7 @@
         "answer": "발톱은 주기적으로 깎아주는 것이 좋습니다. 너무 길어지면 걸을 때 불편함을 줄 수 있고, 심한 경우 발톱이 말려 들어가 피부를 다칠 수 있습니다."
     },
     {
-        "category": "TEETH",
+        "category": "TOOTH",
         "question": "이빨이 자주 아파요.",
         "answer": "치아 문제가 있는 경우 통증이 발생할 수 있습니다. 치석 제거와 같은 치아 관리를 주기적으로 해주고, 필요한 경우 수의사와 상담하세요."
     }
