+++ ttakkeun/ttakkeun/ViewGroup/QnaTab/QnaTipsView.swift
@@ -6,10 +6,25 @@
 //
 import SwiftUI
 
-struct TipsView: View {
+/// Qna탭중 Tips에 대한 뷰
+struct QnaTipsView: View {
     var body: some View {
         VStack {
             /*TipsView 만들자*/
+        }
+    }
+}
+
+//MARK: - Preview
+struct QnaTipsView_Preview: PreviewProvider {
+    
+    static let devices = ["iPhone 11", "iPhone 15 Pro"]
+    
+    static var previews: some View {
+        ForEach(devices, id: \.self) { device in
+            QnaTipsView()
+                .previewDevice(PreviewDevice(rawValue: device))
+                .previewDisplayName(device)
         }
     }
 }
